reporter/requirements.txt
Django==2.1
djangorestframework==3.9.1
psycopg2==2.7.6